+++ ApplicationLibCode/Application/RiaSummaryCurveDefinition.h
@@ -51,7 +51,7 @@ public:
     bool operator<( const RiaSummaryCurveDefinition& other ) const;
 
     // TODO: Consider moving to a separate tools class
-    static void resultValues( const RiaSummaryCurveDefinition& curveDefinition, gsl::not_null<std::vector<double>*> values );
+    static void resultValues( const RiaSummaryCurveDefinition& curveDefinition, gsl::not_null<std::vector<float>*> values );
     static const std::vector<time_t>& timeSteps( const RiaSummaryCurveDefinition& curveDefinition );
 
     QString curveDefinitionText() const;
